TS config was made
